♻️ : #468 - 계획 페이지 로직 분리
